postgres_loader/requirements.txt
python-dotenv==0.18.0
psycopg2-binary==2.9.1
SQLAlchemy==1.4.18
pandas==1.2.4